fabricranker/ml_predictor.py
"""
딥러닝 모델 예측 모듈
실제 환경에서는 TensorFlow, PyTorch 등을 사용하여 구현
"""

import random
import time
from PIL import Image
import os

# 가능한 재질 목록
FABRIC_MATERIALS = [
    "면",
    "폴리에스터",
    "나일론",
    "울",
    "린넨",
    "실크",
    "레이온",
    "아크릴",
    "스판덱스",
    "데님",
    "코튼",
    "캐시미어",
]


def predict_fabric_material(image_path):
    """
    옷 이미지에서 재질을 예측하는 함수

    Args:
        image_path (str): 이미지 파일 경로

    Returns:
        list: [(재질명, 신뢰도), ...] 형태의 상위 3개 결과
    """

    # 실제 구현에서는 여기에 딥러닝 모델 로드 및 예측 코드가 들어갑니다
    # 예시:
    # model = load_model('fabric_classification_model.h5')
    # image = preprocess_image(image_path)
    # predictions = model.predict(image)

    try:
        # 이미지 파일 존재 확인
        if not os.path.exists(image_path):
            raise FileNotFoundError(f"이미지 파일을 찾을 수 없습니다: {image_path}")

        # 이미지 로드 및 기본 검증
        with Image.open(image_path) as img:
            # 이미지 크기 확인
            if img.size[0] < 50 or img.size[1] < 50:
                raise ValueError("이미지가 너무 작습니다.")

        # 모델 예측 시뮬레이션 (실제로는 딥러닝 모델 사용)
        time.sleep(1)  # 모델 처리 시간 시뮬레이션

        # 랜덤하게 3개 재질 선택 및 신뢰도 생성
        selected_materials = random.sample(FABRIC_MATERIALS, 3)

        # 신뢰도는 내림차순으로 정렬 (첫 번째가 가장 높음)
        confidences = sorted(
            [
                random.uniform(0.6, 0.95),  # 60-95% 신뢰도
                random.uniform(0.4, 0.8),  # 40-80% 신뢰도
                random.uniform(0.2, 0.6),  # 20-60% 신뢰도
            ],
            reverse=True,
        )

        # 결과 조합
        results = list(zip(selected_materials, confidences))

        return results

    except Exception as e:
        # 오류 발생 시 기본값 반환
        print(f"예측 중 오류 발생: {str(e)}")
        return [("면", 0.75), ("폴리에스터", 0.65), ("나일론", 0.45)]


def load_real_model():
    """
    실제 딥러닝 모델을 로드하는 함수 (구현 예시)
    """
    # 실제 구현 예시:
    # import tensorflow as tf
    # model = tf.keras.models.load_model('models/fabric_classifier.h5')
    # return model
    pass


def preprocess_image(image_path, target_size=(224, 224)):
    """
    이미지 전처리 함수 (구현 예시)
    """
    # 실제 구현 예시:
    # from tensorflow.keras.preprocessing import image
    # import numpy as np
    #
    # img = image.load_img(image_path, target_size=target_size)
    # img_array = image.img_to_array(img)
    # img_array = np.expand_dims(img_array, axis=0)
    # img_array /= 255.0
    # return img_array
    pass
