requirements.txt
Django>=4.2.0
Pillow>=9.0.0
# 실제 딥러닝 모델 사용 시 추가:
# tensorflow>=2.10.0
# torch>=1.12.0
# torchvision>=0.13.0
# numpy>=1.21.0
# opencv-python>=4.6.0
